requirement.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.7.1
bootstrap4==0.0.5
boto==2.49.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
constantly==15.1.0
coverage==4.5.2
cryptography==2.6.1
csscompressor==0.9.5
cssselect==1.0.3
Cython==0.29.6
dj-database-url==0.5.0
Django==2.2
django-analytical==2.5.0
django-appconf==1.0.2
django-blog-it==0.3.1
django-blogango==0.6.2
django-bootstrap==0.2.4
django-bootstrap4==0.0.7
django-classy-tags==0.8.0
django-cms==3.6.0
django-compressor==2.2
django-contrib-comments==1.9.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-djconfig==0.8.0
django-eventtools==0.9.11
django-filebrowser-no-grappelli==3.7.7
django-filer==1.4.4
django-formtools==2.1
django-haystack==2.8.1
django-heroku==0.3.1
django-infinite-scroll-pagination==0.2.0
django-js-asset==1.2.2
django-mailer==1.2.5
django-mptt==0.9.1
-e git://github.com/dokterbob/django-newsletter.git@30c3ec3f3bea93b22c6640aed40ee2c8040a4f46#egg=django_newsletter
django-polymorphic==2.0.3
django-ranged-response==0.2.0
django-registration-redux==2.5
django-request-provider==1.0.2
django-sekizai==0.10.0
django-simple-captcha==0.5.10
django-simple-pagination==1.3
django-spirit==0.7.0
django-storages==1.7.1
django-suit==0.2.26
django-suit-ckeditor==0.0.2
django-summernote==0.8.11.4
django-taggit==0.24.0
django-tinymce==2.8.0
django-treebeard==4.3
djangocms-admin-style==1.3.0
docopt==0.6.2
easy-thumbnails==2.6
entrypoints==0.3
flake8==3.7.7
gunicorn==19.9.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
Jinja2==2.10
ldif3==3.1.1
lockfile==0.12.2
lxml==4.3.2
MarkupSafe==1.1.1
mccabe==0.6.1
microurl==0.1.1
mistune==0.8.4
mock==2.0.0
olefile==0.46
paginate==0.5.6
parsel==1.5.1
pbr==5.1.3
pep8==1.7.1
Pillow==5.4.1
pipdeptree==0.13.2
pipreqs==0.4.9
psycopg2==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
pyflakes==2.1.1
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
PyPDF2==1.26.0
python-card-me==0.9.3
python-dateutil==2.8.0
pytils==0.3
pytz==2018.9
qrcode==6.1
queuelib==1.5.0
rcssmin==1.0.6
reportlab==3.5.13
requests==2.21.0
rjsmin==1.0.12
Scrapy==1.6.0
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
sorl-thumbnail==12.5.0
soupsieve==1.8
sqlparse==0.2.4
surlex==0.2.0
Twisted==18.9.0
uni-slugify==0.1.4
unicodecsv==0.14.1
Unidecode==1.0.23
urllib3==1.24.1
virtualenv==16.4.3
virtualenv-clone==0.5.1
w3lib==1.20.0
webencodings==0.5.1
whitenoise==4.1.2
Whoosh==2.7.4
xhtml2pdf==0.2.3
yarg==0.1.9
zope.interface==4.6.0
